requirements.txt
Django==3.2.12
djangorestframework==3.13.1
django-filter==21.1
django-cors-headers==3.10.0
psycopg2-binary==2.9.3
Pillow==9.0.1
python-decouple==3.6
gunicorn==20.1.0
Faker==18.13.0
